+++ django-data/image/cryoweb/tests/test_helpers.py
@@ -363,8 +363,10 @@ class UploadCryoweb(
 
             # check async message called
             message = 'Error'
-            notification_message = ('Error in importing data: Test '
-                                    'upload failed')
+            notification_message = (
+                "Error in importing data: Test upload failed\nIs "
+                "'cryoweb_test_data_only.sql' a valid cryoweb dump file?"
+            )
 
             self.check_message(message, notification_message)